feat: split endpoints
